+++ src/idc/filter/_split_records.py
@@ -12,10 +12,10 @@ META_SPLIT = "split"
 
 class SplitRecords(Filter):
     """
-    Splits the incoming records into the specified split ratios by setting the 'split' meta-data value. Also stores the split names in the current session.
+    Splits the incoming records into the specified split ratios by setting the 'split' meta-data value.
     """
 
-    def __init__(self, split_ratios: List[int] = None, split_names: List[str] = None,
+    def __init__(self, split_ratios: List[int] = None, split_names: List[str] = None, field: str = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
         Initializes the filter.
@@ -24,6 +24,8 @@ class SplitRecords(Filter):
         :type split_ratios: list
         :param split_names: the names for the splits (will be stored in the meta-data)
         :type split_names: list
+        :param field: the metadata field to store the split in
+        :type field: str
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -32,6 +34,7 @@ class SplitRecords(Filter):
         super().__init__(logger_name=logger_name, logging_level=logging_level)
         self.split_ratios = split_ratios
         self.split_names = split_names
+        self.field = field
         self._splitter = None
 
     def name(self) -> str:
@@ -50,7 +53,7 @@ class SplitRecords(Filter):
         :return: the description
         :rtype: str
         """
-        return "Splits the incoming records into the specified split ratios by setting the '%s' meta-data value. Also stores the split names in the current session." % META_SPLIT
+        return "Splits the incoming records into the specified split ratios by setting a meta-data value (default: %s)." % META_SPLIT
 
     def accepts(self) -> List:
         """
@@ -79,7 +82,8 @@ class SplitRecords(Filter):
         """
         parser = super()._create_argparser()
         parser.add_argument("-r", "--split_ratios", type=int, default=None, help="The split ratios to use for generating the splits (must sum up to 100)", nargs="+")
-        parser.add_argument("-n", "--split_names", type=str, default=None, help="The split names to use for the generated splits, get stored in the meta-data under the key '" + META_SPLIT + "'.", nargs="+")
+        parser.add_argument("-n", "--split_names", type=str, default=None, help="The split names to use for the generated splits, get stored in the meta-data under the specified key.", nargs="+")
+        parser.add_argument("-f", "--field", type=str, help="The meta-data field to use for storing the split", default=META_SPLIT, required=False)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -92,6 +96,7 @@ class SplitRecords(Filter):
         super()._apply_args(ns)
         self.split_ratios = ns.split_ratios[:]
         self.split_names = ns.split_names[:]
+        self.field = ns.field
 
     def initialize(self):
         """
@@ -101,6 +106,8 @@ class SplitRecords(Filter):
 
         self._splitter = Splitter(self.split_ratios, self.split_names)
         self._splitter.initialize()
+        if (self.field is None) or (len(self.field.strip()) == 0):
+            self.field = META_SPLIT
 
     def _output_stats(self):
         """
@@ -135,7 +142,7 @@ class SplitRecords(Filter):
                     item.set_metadata(meta)
 
             # find split according to schedule
-            meta[META_SPLIT] = self._splitter.next()
+            meta[self.field] = self._splitter.next()
             result.append(item)
 
         return flatten_list(result)