+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - added `dims-to-metadata` filter for transferring image dimension to metadata
 - centralized comparison code in `idc.api._comparison` (available via `idc.api`)
+- `split-records` filter now allows specifying the meta-data field in which to store the split name
 
 
 0.0.13 (2025-07-15)